optimization

+++ cmd/mcp_dbmem/flag/database.go
@@ -9,7 +9,7 @@ import (
 func Database(cmd *cobra.Command, values config.Values) {
 	cmd.PersistentFlags().String(config.Keys.DBType, values.DBType, usage.DBType)
 	cmd.PersistentFlags().String(config.Keys.DBAddress, values.DBAddress, usage.DBAddress)
-	cmd.PersistentFlags().Int(config.Keys.DBPort, values.DBPort, usage.DBPort)
+	cmd.PersistentFlags().Uint16(config.Keys.DBPort, values.DBPort, usage.DBPort)
 	cmd.PersistentFlags().String(config.Keys.DBUser, values.DBUser, usage.DBUser)
 	cmd.PersistentFlags().String(config.Keys.DBPassword, values.DBPassword, usage.DBPassword)
 	cmd.PersistentFlags().String(config.Keys.DBDatabase, values.DBDatabase, usage.DBDatabase)
